Changes made considering differences between the latest commit and viniciusBranch: TrainFunctions module updated.
Scores Functions separated and moved to packages organized with respect to differences in implementation and end use.

CNN functions moved to a separated module: ConvolutionalNeuralNetworks

Pushed CNN train scripts
