+++ Functions/TrainFunctions.py
@@ -4,11 +4,7 @@
 
     Description: This file contains functions used for Neural Network training.
 """
-import csv
 import os
-import pickle
-import warnings
-from abc import abstractmethod
 from warnings import warn
 import pandas as pd
 import numpy as np
@@ -16,23 +12,20 @@ import time
 
 from sklearn.externals import joblib
 from sklearn import preprocessing
-from sklearn import metrics
 
 from keras.models import Sequential
 from keras.layers.core import Dense, Activation
-from keras.optimizers import Adam, SGD
+from keras.optimizers import Adam
 import keras.callbacks as callbacks
 from keras.utils import np_utils
 from keras.models import load_model
-from keras import backend as K, Model
+from keras import backend as K
 
 import multiprocessing
 
-from Functions import TrainParameters as trnparams, SystemIO, DataHandler
-from Functions.DataHandler import lofar2image
-from Functions.SystemIO import save, load
-from Functions.TrainParameters import TrnParamsConvolutional
-from Functions.TrainPaths import ConvolutionPaths, ModelPaths
+import Functions.NpUtils.DataTransformation
+from Functions import TrainParameters as trnparams, SystemIO, CrossValidation
+from Functions.ConvolutionalNeuralNetworks import ConvolutionPaths
 
 num_process = multiprocessing.cpu_count()
 
@@ -40,7 +33,6 @@ num_process = multiprocessing.cpu_count()
     Function used to do a Fine Tuning in Stacked Auto Encoder for Classification of the data
     hidden_neurons contains the number of neurons in the sequence: [FirstLayer, SecondLayer, ... ]
 '''
-
 
 def SAEClassificationTrainFunction(data=None, trgt=None, ifold=0,
                                    n_folds=2, hidden_neurons=[],
@@ -54,7 +46,7 @@ def SAEClassificationTrainFunction(data=None, trgt=None, ifold=0,
 
     trgt_sparse = np_utils.to_categorical(trgt.astype(int))
     # load or create cross validation ids
-    CVO = trnparams.ClassificationFolds(folder=save_path, n_folds=n_folds, trgt=trgt, dev=dev)
+    CVO = trnparams.ClassificationFolds(folder=save_path,n_folds=n_folds,trgt=trgt,dev=dev)
 
     n_folds = len(CVO)
     n_inits = trn_params.params['n_inits']
@@ -69,33 +61,34 @@ def SAEClassificationTrainFunction(data=None, trgt=None, ifold=0,
     for ineuron in hidden_neurons:
         neurons_str = neurons_str + 'x' + str(ineuron)
     neurons_str = neurons_str + 'x' + str(trgt_sparse.shape[1])
-    model_str = '%s/%s/Classification_(%s)_%s_%i_folds_%s' % (save_path, analysis_str,
-                                                              neurons_str,
-                                                              prefix_str, n_folds,
-                                                              params_str)
+    model_str = '%s/%s/Classification_(%s)_%s_%i_folds_%s'%(save_path,analysis_str,
+                                                                     neurons_str,
+                                                                     prefix_str, n_folds,
+                                                                     params_str)
+
 
     if not dev:
-        file_name = '%s_fold_%i_model.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model.h5'%(model_str,ifold)
         if os.path.exists(file_name):
             if trn_params.params['verbose']:
-                print 'File %s exists' % (file_name)
+                print 'File %s exists'%(file_name)
             return 0
     else:
-        file_name = '%s_fold_%i_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model_dev.h5'%(model_str,ifold)
         if os.path.exists(file_name):
             if trn_params.params['verbose']:
-                print 'File %s exists' % (file_name)
+                print 'File %s exists'%(file_name)
             return 0
 
     train_id, test_id = CVO[ifold]
 
-    # normalize data based in train set
+    #normalize data based in train set
     if trn_params.params['norm'] == 'mapstd':
-        scaler = preprocessing.StandardScaler().fit(data[train_id, :])
+        scaler = preprocessing.StandardScaler().fit(data[train_id,:])
     elif trn_params.params['norm'] == 'mapstd_rob':
-        scaler = preprocessing.RobustScaler().fit(data[train_id, :])
+        scaler = preprocessing.RobustScaler().fit(data[train_id,:])
     elif trn_params.params['norm'] == 'mapminmax':
-        scaler = preprocessing.MinMaxScaler().fit(data[train_id, :])
+        scaler = preprocessing.MinMaxScaler().fit(data[train_id,:])
 
     norm_data = scaler.transform(data)
 
@@ -106,39 +99,39 @@ def SAEClassificationTrainFunction(data=None, trgt=None, ifold=0,
     trn_desc = {}
 
     for i_init in range(n_inits):
-        print 'Fold %i of %i Folds -  Init %i of %i Inits' % (ifold + 1,
-                                                              n_folds,
-                                                              i_init + 1,
-                                                              n_inits)
+        print 'Fold %i of %i Folds -  Init %i of %i Inits'%(ifold+1,
+                                                            n_folds,
+                                                            i_init+1,
+                                                            n_inits)
 
         # Get the weights of first layer
         layer = 1
         neurons_str = str(data.shape[1])
         for ineuron in hidden_neurons[:layer]:
             neurons_str = neurons_str + 'x' + str(ineuron)
-        previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                  prefix_str,
-                                                                  n_folds,
-                                                                  params_str,
-                                                                  neurons_str)
+        previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                prefix_str,
+                                                                n_folds,
+                                                                params_str,
+                                                                neurons_str)
 
         if not dev:
-            file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+            file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
         else:
-            file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+            file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
         if not os.path.exists(file_name):
             def trainFold(ifold):
                 ineuron = hidden_neurons[1]
                 return StackedAutoEncoderTrainFunction(data=all_data,
-                                                       trgt=all_data,
-                                                       ifold=ifold,
-                                                       n_folds=n_folds,
-                                                       n_neurons=ineuron,
-                                                       trn_params=trn_params,
-                                                       save_path=results_path,
-                                                       layer=layer,
-                                                       hidden_neurons=hidden_neurons[:layer],
-                                                       dev=dev)
+                                        trgt=all_data,
+                                        ifold=ifold,
+                                        n_folds=n_folds,
+                                        n_neurons=ineuron,
+                                        trn_params=trn_params,
+                                        save_path=results_path,
+                                        layer=layer,
+                                        hidden_neurons = hidden_neurons[:layer],
+                                        dev=dev)
 
             p = Pool(processes=num_processes)
             folds = range(len(CVO))
@@ -154,35 +147,35 @@ def SAEClassificationTrainFunction(data=None, trgt=None, ifold=0,
         model.add(Dense(hidden_neurons[0], input_dim=norm_data.shape[1], weights=first_layer_weights, trainable=True))
         model.add(Activation(trn_params.params['hidden_activation']))
         # Add second hidden layer
-        if (len(hidden_neurons) > 1):
+        if(len(hidden_neurons) > 1):
             # Get the weights of second layer
             layer = 2
             neurons_str = str(data.shape[1])
             for ineuron in hidden_neurons[:layer]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
 
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
             if not os.path.exists(file_name):
                 def trainFold(ifold):
                     ineuron = hidden_neurons[1]
                     return StackedAutoEncoderTrainFunction(data=all_data,
-                                                           trgt=all_data,
-                                                           ifold=ifold,
-                                                           n_folds=n_folds,
-                                                           n_neurons=ineuron,
-                                                           trn_params=trn_params,
-                                                           save_path=results_path,
-                                                           layer=layer,
-                                                           hidden_neurons=hidden_neurons[:layer],
-                                                           dev=dev)
+                                            trgt=all_data,
+                                            ifold=ifold,
+                                            n_folds=n_folds,
+                                            n_neurons=ineuron,
+                                            trn_params=trn_params,
+                                            save_path=results_path,
+                                            layer=layer,
+                                            hidden_neurons = hidden_neurons[:layer],
+                                            dev=dev)
 
                 p = Pool(processes=num_processes)
                 folds = range(len(CVO))
@@ -197,35 +190,35 @@ def SAEClassificationTrainFunction(data=None, trgt=None, ifold=0,
             model.add(Dense(hidden_neurons[1], weights=second_layer_weights, trainable=True))
             model.add(Activation(trn_params.params['hidden_activation']))
         # Add third hidden layer
-        if (len(hidden_neurons) > 2):
-            # Get the weights of third layer
+        if(len(hidden_neurons) > 2):
+             # Get the weights of third layer
             layer = 3
             neurons_str = str(data.shape[1])
             for ineuron in hidden_neurons[:layer]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
 
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
             if not os.path.exists(file_name):
                 def trainFold(ifold):
                     ineuron = hidden_neurons[1]
                     return StackedAutoEncoderTrainFunction(data=all_data,
-                                                           trgt=all_data,
-                                                           ifold=ifold,
-                                                           n_folds=n_folds,
-                                                           n_neurons=ineuron,
-                                                           trn_params=trn_params,
-                                                           save_path=results_path,
-                                                           layer=layer,
-                                                           hidden_neurons=hidden_neurons[:layer],
-                                                           dev=dev)
+                                            trgt=all_data,
+                                            ifold=ifold,
+                                            n_folds=n_folds,
+                                            n_neurons=ineuron,
+                                            trn_params=trn_params,
+                                            save_path=results_path,
+                                            layer=layer,
+                                            hidden_neurons = hidden_neurons[:layer],
+                                            dev=dev)
 
                 p = Pool(processes=num_processes)
                 folds = range(len(CVO))
@@ -240,35 +233,35 @@ def SAEClassificationTrainFunction(data=None, trgt=None, ifold=0,
             model.add(Dense(hidden_neurons[2], weights=third_layer_weights, trainable=True))
             model.add(Activation(trn_params.params['hidden_activation']))
         # Add fourth hidden layer
-        if (len(hidden_neurons) > 3):
-            # Get the weights of fourth layer
+        if(len(hidden_neurons) > 3):
+             # Get the weights of fourth layer
             layer = 4
             neurons_str = str(data.shape[1])
             for ineuron in hidden_neurons[:layer]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
 
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
             if not os.path.exists(file_name):
                 def trainFold(ifold):
                     ineuron = hidden_neurons[1]
                     return StackedAutoEncoderTrainFunction(data=all_data,
-                                                           trgt=all_data,
-                                                           ifold=ifold,
-                                                           n_folds=n_folds,
-                                                           n_neurons=ineuron,
-                                                           trn_params=trn_params,
-                                                           save_path=results_path,
-                                                           layer=layer,
-                                                           hidden_neurons=hidden_neurons[:layer],
-                                                           dev=dev)
+                                            trgt=all_data,
+                                            ifold=ifold,
+                                            n_folds=n_folds,
+                                            n_neurons=ineuron,
+                                            trn_params=trn_params,
+                                            save_path=results_path,
+                                            layer=layer,
+                                            hidden_neurons = hidden_neurons[:layer],
+                                            dev=dev)
 
                 p = Pool(processes=num_processes)
                 folds = range(len(CVO))
@@ -321,31 +314,29 @@ def SAEClassificationTrainFunction(data=None, trgt=None, ifold=0,
 
     # save model
     if not dev:
-        file_name = '%s_fold_%i_model.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model.h5'%(model_str,ifold)
         classifier.save(file_name)
-        file_name = '%s_fold_%i_trn_desc.jbl' % (model_str, ifold)
-        joblib.dump([trn_desc], file_name, compress=9)
+        file_name = '%s_fold_%i_trn_desc.jbl'%(model_str,ifold)
+        joblib.dump([trn_desc],file_name,compress=9)
     else:
-        file_name = '%s_fold_%i_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model_dev.h5'%(model_str,ifold)
         classifier.save(file_name)
-        file_name = '%s_fold_%i_trn_desc_dev.jbl' % (model_str, ifold)
+        file_name = '%s_fold_%i_trn_desc_dev.jbl'%(model_str,ifold)
         print file_name
-        joblib.dump([trn_desc], file_name, compress=9)
-
+        joblib.dump([trn_desc],file_name,compress=9)
 
 '''
     Function used to perform the layerwise algorithm to train the SAE
 '''
-
-
 def StackedAutoEncoderTrainFunction(data=None, trgt=None,
                                     ifold=0, n_folds=2, n_neurons=10,
-                                    trn_params=None, save_path='', dev=False, layer=1, hidden_neurons=[400]):
+                                    trn_params=None, save_path='', dev=False, layer = 1, hidden_neurons = [400]):
+
     if n_neurons == 0:
         n_neurons = 1
 
     # load or create cross validation ids
-    CVO = trnparams.ClassificationFolds(folder=save_path, n_folds=n_folds, trgt=trgt, dev=dev)
+    CVO = trnparams.ClassificationFolds(folder=save_path,n_folds=n_folds,trgt=trgt,dev=dev)
 
     n_folds = len(CVO)
     n_inits = trn_params.params['n_inits']
@@ -357,41 +348,42 @@ def StackedAutoEncoderTrainFunction(data=None, trgt=None,
 
     if layer == 1:
         neurons_str = str(data.shape[1]) + 'x' + str(n_neurons)
-        model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                         prefix_str, n_folds,
-                                                         params_str, neurons_str)
+        model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                          prefix_str,n_folds,
+                                                          params_str,neurons_str)
 
     else:
         neurons_str = str(data.shape[1])
         for ineuron in hidden_neurons[:layer]:
             neurons_str = neurons_str + 'x' + str(ineuron)
         neurons_str = neurons_str + 'x' + str(n_neurons)
-        model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                         prefix_str, n_folds,
-                                                         params_str, neurons_str)
+        model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                prefix_str,n_folds,
+                                                                params_str,neurons_str)
+
 
     if not dev:
-        file_name = '%s_fold_%i_model.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model.h5'%(model_str,ifold)
         if os.path.exists(file_name):
             if trn_params.params['verbose']:
-                print 'File %s exists' % (file_name)
+                print 'File %s exists'%(file_name)
             return 0
     else:
-        file_name = '%s_fold_%i_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model_dev.h5'%(model_str,ifold)
         if os.path.exists(file_name):
             if trn_params.params['verbose']:
-                print 'File %s exists' % (file_name)
+                print 'File %s exists'%(file_name)
             return 0
 
     train_id, test_id = CVO[ifold]
 
-    # normalize data based in train set
+    #normalize data based in train set
     if trn_params.params['norm'] == 'mapstd':
-        scaler = preprocessing.StandardScaler().fit(data[train_id, :])
+        scaler = preprocessing.StandardScaler().fit(data[train_id,:])
     elif trn_params.params['norm'] == 'mapstd_rob':
-        scaler = preprocessing.RobustScaler().fit(data[train_id, :])
+        scaler = preprocessing.RobustScaler().fit(data[train_id,:])
     elif trn_params.params['norm'] == 'mapminmax':
-        scaler = preprocessing.MinMaxScaler().fit(data[train_id, :])
+        scaler = preprocessing.MinMaxScaler().fit(data[train_id,:])
 
     norm_data = scaler.transform(data)
     norm_trgt = norm_data
@@ -403,11 +395,11 @@ def StackedAutoEncoderTrainFunction(data=None, trgt=None,
     trn_desc = {}
 
     for i_init in range(n_inits):
-        print 'Neuron: %i - Fold %i of %i Folds -  Init %i of %i Inits' % (n_neurons,
-                                                                           ifold + 1,
-                                                                           n_folds,
-                                                                           i_init + 1,
-                                                                           n_inits)
+        print 'Neuron: %i - Fold %i of %i Folds -  Init %i of %i Inits'%(n_neurons,
+                                                                         ifold+1,
+                                                                         n_folds,
+                                                                         i_init+1,
+                                                                         n_inits)
         if layer == 1:
             model = Sequential()
             model.add(Dense(n_neurons, input_dim=data.shape[1], init="uniform"))
@@ -417,17 +409,17 @@ def StackedAutoEncoderTrainFunction(data=None, trgt=None,
         if layer == 2:
             # Get the projection of the data from previous layer
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 1]:
+            for ineuron in hidden_neurons[:layer-1]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             ################################
             # Check if previous model was trained
@@ -449,44 +441,44 @@ def StackedAutoEncoderTrainFunction(data=None, trgt=None,
         if layer == 3:
             # Load first layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 2]:
+            for ineuron in hidden_neurons[:layer-2]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             first_layer_model = load_model(file_name)
 
             get_first_layer_output = K.function([first_layer_model.layers[0].input],
-                                                [first_layer_model.layers[1].output])
+                                          [first_layer_model.layers[1].output])
 
             # Projection of first layer
             first_proj_all_data = get_first_layer_output([norm_data])[0]
 
             # Load second layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 1]:
+            for ineuron in hidden_neurons[:layer-1]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             second_layer_model = load_model(file_name)
 
             get_second_layer_output = K.function([second_layer_model.layers[0].input],
-                                                 [second_layer_model.layers[1].output])
+                                          [second_layer_model.layers[1].output])
 
             # Projection of second layer
             second_proj_all_data = get_second_layer_output([first_proj_all_data])[0]
@@ -501,66 +493,66 @@ def StackedAutoEncoderTrainFunction(data=None, trgt=None,
         if layer == 4:
             # Load first layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 3]:
+            for ineuron in hidden_neurons[:layer-3]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             first_layer_model = load_model(file_name)
 
             get_first_layer_output = K.function([first_layer_model.layers[0].input],
-                                                [first_layer_model.layers[1].output])
+                                          [first_layer_model.layers[1].output])
 
             # Projection of first layer
             first_proj_all_data = get_first_layer_output([norm_data])[0]
 
             # Load second layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 2]:
+            for ineuron in hidden_neurons[:layer-2]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             second_layer_model = load_model(file_name)
 
             get_second_layer_output = K.function([second_layer_model.layers[0].input],
-                                                 [second_layer_model.layers[1].output])
+                                          [second_layer_model.layers[1].output])
 
             # Projection of second layer
             second_proj_all_data = get_second_layer_output([first_proj_all_data])[0]
 
             # Load third layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 1]:
+            for ineuron in hidden_neurons[:layer-1]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             third_layer_model = load_model(file_name)
 
             get_third_layer_output = K.function([third_layer_model.layers[0].input],
-                                                [third_layer_model.layers[1].output])
+                                          [third_layer_model.layers[1].output])
 
             # Projection of third layer
             third_proj_all_data = get_third_layer_output([second_proj_all_data])[0]
@@ -571,7 +563,7 @@ def StackedAutoEncoderTrainFunction(data=None, trgt=None,
             model.add(Dense(third_proj_all_data.shape[1], init="uniform"))
             model.add(Activation(trn_params.params['output_activation']))
             norm_data = third_proj_all_data
-        # end if layer == 4
+        #end if layer == 4
 
         # Optimizer
         adam = Adam(lr=trn_params.params['learning_rate'],
@@ -609,27 +601,26 @@ def StackedAutoEncoderTrainFunction(data=None, trgt=None,
 
     # save model
     if not dev:
-        file_name = '%s_fold_%i_model.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model.h5'%(model_str,ifold)
         classifier.save(file_name)
-        file_name = '%s_fold_%i_trn_desc.jbl' % (model_str, ifold)
-        joblib.dump([trn_desc], file_name, compress=9)
+        file_name = '%s_fold_%i_trn_desc.jbl'%(model_str,ifold)
+        joblib.dump([trn_desc],file_name,compress=9)
     else:
-        file_name = '%s_fold_%i_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model_dev.h5'%(model_str,ifold)
         classifier.save(file_name)
-        file_name = '%s_fold_%i_trn_desc_dev.jbl' % (model_str, ifold)
-        joblib.dump([trn_desc], file_name, compress=9)
-
+        file_name = '%s_fold_%i_trn_desc_dev.jbl'%(model_str,ifold)
+        joblib.dump([trn_desc],file_name,compress=9)
 
 '''
     Function used to perform the training of a Classifier with one hidden layer.
     Topology(in terms of neurons):
         InputDim(400) x n_neurons x OutputDim(Num of Classes)
 '''
-
-
 def NeuralTrainFunction(data=None, trgt=None,
                         ifold=0, n_folds=2, n_neurons=10,
                         trn_params=None, save_path='', dev=False):
+
+
     if n_neurons == 0:
         n_neurons = 1
 
@@ -637,7 +628,7 @@ def NeuralTrainFunction(data=None, trgt=None,
     trgt_sparse = np_utils.to_categorical(trgt.astype(int))
 
     # load or create cross validation ids
-    CVO = trnparams.ClassificationFolds(folder=save_path, n_folds=n_folds, trgt=trgt, dev=dev)
+    CVO = trnparams.ClassificationFolds(folder=save_path,n_folds=n_folds,trgt=trgt,dev=dev)
 
     n_folds = len(CVO)
     n_inits = trn_params.params['n_inits']
@@ -647,32 +638,33 @@ def NeuralTrainFunction(data=None, trgt=None,
     analysis_str = 'NeuralNetwork'
     prefix_str = 'RawData'
 
-    model_str = '%s/%s/%s_%i_folds_%s_%i_neurons' % (save_path, analysis_str,
-                                                     prefix_str, n_folds,
-                                                     params_str, n_neurons)
+    model_str = '%s/%s/%s_%i_folds_%s_%i_neurons'%(save_path,analysis_str,
+                                                   prefix_str,n_folds,
+                                                   params_str,n_neurons)
+
 
     if not dev:
-        file_name = '%s_fold_%i_model.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model.h5'%(model_str,ifold)
         if os.path.exists(file_name):
             if trn_params.params['verbose']:
-                print 'File %s exists' % (file_name)
+                print 'File %s exists'%(file_name)
             return 0
     else:
-        file_name = '%s_fold_%i_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model_dev.h5'%(model_str,ifold)
         if os.path.exists(file_name):
             if trn_params.params['verbose']:
-                print 'File %s exists' % (file_name)
+                print 'File %s exists'%(file_name)
             return 0
 
     train_id, test_id = CVO[ifold]
 
-    # normalize data based in train set
+    #normalize data based in train set
     if trn_params.params['norm'] == 'mapstd':
-        scaler = preprocessing.StandardScaler().fit(data[train_id, :])
+        scaler = preprocessing.StandardScaler().fit(data[train_id,:])
     elif trn_params.params['norm'] == 'mapstd_rob':
-        scaler = preprocessing.RobustScaler().fit(data[train_id, :])
+        scaler = preprocessing.RobustScaler().fit(data[train_id,:])
     elif trn_params.params['norm'] == 'mapminmax':
-        scaler = preprocessing.MinMaxScaler().fit(data[train_id, :])
+        scaler = preprocessing.MinMaxScaler().fit(data[train_id,:])
 
     norm_data = scaler.transform(data)
 
@@ -683,11 +675,11 @@ def NeuralTrainFunction(data=None, trgt=None,
     trn_desc = {}
 
     for i_init in range(n_inits):
-        print 'Neuron: %i - Fold %i of %i Folds -  Init %i of %i Inits' % (n_neurons,
-                                                                           ifold + 1,
-                                                                           n_folds,
-                                                                           i_init + 1,
-                                                                           n_inits)
+        print 'Neuron: %i - Fold %i of %i Folds -  Init %i of %i Inits'%(n_neurons,
+                                                                         ifold+1,
+                                                                         n_folds,
+                                                                         i_init+1,
+                                                                         n_inits)
         model = Sequential()
         model.add(Dense(n_neurons, input_dim=data.shape[1], init="uniform"))
         model.add(Activation(trn_params.params['hidden_activation']))
@@ -728,51 +720,49 @@ def NeuralTrainFunction(data=None, trgt=None,
 
     # save model
     if not dev:
-        file_name = '%s_fold_%i_model.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model.h5'%(model_str,ifold)
         classifier.save(file_name)
-        file_name = '%s_fold_%i_trn_desc.jbl' % (model_str, ifold)
-        joblib.dump([trn_desc], file_name, compress=9)
+        file_name = '%s_fold_%i_trn_desc.jbl'%(model_str,ifold)
+        joblib.dump([trn_desc],file_name,compress=9)
     else:
-        file_name = '%s_fold_%i_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model_dev.h5'%(model_str,ifold)
         classifier.save(file_name)
-        file_name = '%s_fold_%i_trn_desc_dev.jbl' % (model_str, ifold)
-        joblib.dump([trn_desc], file_name, compress=9)
-
+        file_name = '%s_fold_%i_trn_desc_dev.jbl'%(model_str,ifold)
+        joblib.dump([trn_desc],file_name,compress=9)
 
 '''
     Function used to perform the training of a Classifier and Novelty Detector
 '''
-
-
 def NNNoveltyTrainFunction(data=None, trgt=None, inovelty=0,
                            ifold=0, n_folds=2, n_neurons=1,
                            trn_params=None, save_path='',
                            verbose=False, dev=False):
+
     # turn targets in sparse mode
     trgt_sparse = np_utils.to_categorical(trgt.astype(int))
 
     # load or create cross validation ids
-    CVO = trnparams.NoveltyDetectionFolds(folder=save_path, n_folds=n_folds, trgt=trgt, dev=dev)
+    CVO = trnparams.NoveltyDetectionFolds(folder=save_path,n_folds=n_folds,trgt=trgt,dev=dev)
 
     if n_neurons == 0:
         n_neurons = 1
 
     n_folds = len(CVO[0])
     n_inits = trn_params.params['n_inits']
-    model_prefix_str = 'RawData_%i_novelty' % (inovelty)
+    model_prefix_str = 'RawData_%i_novelty'%(inovelty)
     analysis_path = 'NeuralNetwork'
 
     params_str = trn_params.get_params_str()
 
-    model_str = '%s/%s/%s_%i_folds_%s_%i_neurons' % (save_path, analysis_path,
-                                                     model_prefix_str,
-                                                     n_folds,
-                                                     params_str,
-                                                     n_neurons)
+    model_str = '%s/%s/%s_%i_folds_%s_%i_neurons'%(save_path,analysis_path,
+                                                   model_prefix_str,
+                                                   n_folds,
+                                                   params_str,
+                                                   n_neurons)
     if not dev:
-        file_name = '%s_%i_fold_model.h5' % (model_str, ifold)
+        file_name = '%s_%i_fold_model.h5'%(model_str,ifold)
     else:
-        file_name = '%s_%i_fold_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_%i_fold_model_dev.h5'%(model_str,ifold)
 
     if trn_params.params['verbose']:
         print file_name
@@ -790,11 +780,11 @@ def NNNoveltyTrainFunction(data=None, trgt=None, inovelty=0,
 
         # normalize data based in train set
         if trn_params.params['norm'] == 'mapstd':
-            scaler = preprocessing.StandardScaler().fit(data[train_id, :])
+            scaler = preprocessing.StandardScaler().fit(data[train_id,:])
         elif trn_params.params['norm'] == 'mapstd_rob':
-            scaler = preprocessing.RobustScaler().fit(data[train_id, :])
+            scaler = preprocessing.RobustScaler().fit(data[train_id,:])
         elif trn_params.params['norm'] == 'mapminmax':
-            scaler = preprocessing.MinMaxScaler().fit(data[train_id, :])
+            scaler = preprocessing.MinMaxScaler().fit(data[train_id,:])
 
         norm_data = scaler.transform(data)
 
@@ -802,11 +792,11 @@ def NNNoveltyTrainFunction(data=None, trgt=None, inovelty=0,
         best_loss = 999
 
         for i_init in range(n_inits):
-            print 'Neuron: %i - Fold %i of %i Folds -  Init %i of %i Inits' % (n_neurons,
-                                                                               ifold + 1,
-                                                                               n_folds,
-                                                                               i_init + 1,
-                                                                               n_inits)
+            print 'Neuron: %i - Fold %i of %i Folds -  Init %i of %i Inits'%(n_neurons,
+                                                                             ifold+1,
+                                                                             n_folds,
+                                                                             i_init+1,
+                                                                             n_inits)
             model = Sequential()
             model.add(Dense(n_neurons, input_dim=data.shape[1], init="uniform"))
             model.add(Activation(trn_params.params['hidden_activation']))
@@ -814,9 +804,9 @@ def NNNoveltyTrainFunction(data=None, trgt=None, inovelty=0,
             model.add(Activation(trn_params.params['output_activation']))
 
             adam = Adam(lr=trn_params.params['learning_rate'],
-                        beta_1=trn_params.params['beta_1'],
-                        beta_2=trn_params.params['beta_2'],
-                        epsilon=trn_params.params['epsilon'])
+                    beta_1=trn_params.params['beta_1'],
+                    beta_2=trn_params.params['beta_2'],
+                    epsilon=trn_params.params['epsilon'])
 
             model.compile(loss='mean_squared_error',
                           optimizer=adam,
@@ -848,38 +838,38 @@ def NNNoveltyTrainFunction(data=None, trgt=None, inovelty=0,
 
         # save model
         if not dev:
-            file_name = '%s_%i_fold_model.h5' % (model_str, ifold)
+            file_name = '%s_%i_fold_model.h5'%(model_str,ifold)
             classifier.save(file_name)
-            file_name = '%s_%i_fold_trn_desc.jbl' % (model_str, ifold)
-            joblib.dump([trn_desc], file_name, compress=9)
+            file_name = '%s_%i_fold_trn_desc.jbl'%(model_str,ifold)
+            joblib.dump([trn_desc],file_name,compress=9)
         else:
-            file_name = '%s_%i_fold_model_dev.h5' % (model_str, ifold)
+            file_name = '%s_%i_fold_model_dev.h5'%(model_str,ifold)
             classifier.save(file_name)
-            file_name = '%s_%i_fold_trn_desc_dev.jbl' % (model_str, ifold)
-            joblib.dump([trn_desc], file_name, compress=9)
+            file_name = '%s_%i_fold_trn_desc_dev.jbl'%(model_str,ifold)
+            joblib.dump([trn_desc],file_name,compress=9)
     else:
         # The model has already been trained, so load the files
         if verbose:
             print 'Load Model'
         classifier = Sequential()
         if not dev:
-            file_name = '%s_%i_fold_model.h5' % (model_str, ifold)
+            file_name = '%s_%i_fold_model.h5'%(model_str,ifold)
         else:
-            file_name = '%s_%i_fold_model_dev.h5' % (model_str, ifold)
+            file_name = '%s_%i_fold_model_dev.h5'%(model_str,ifold)
         classifier = load_model(file_name)
 
         if not dev:
-            file_name = '%s_%i_fold_trn_desc.jbl' % (model_str, ifold)
+            file_name = '%s_%i_fold_trn_desc.jbl'%(model_str,ifold)
         else:
-            file_name = '%s_%i_fold_trn_desc_dev.jbl' % (model_str, ifold)
+            file_name = '%s_%i_fold_trn_desc_dev.jbl'%(model_str,ifold)
         [trn_desc] = joblib.load(file_name)
 
-    return [classifier, trn_desc]
+    return [classifier,trn_desc]
 
 
-def SAENoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2, n_neurons=1, layer=1,
-                            hidden_neurons=[400],
+def SAENoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2, n_neurons=1, layer = 1, hidden_neurons = [400],
                             trn_params=None, save_path='', verbose=False, dev=False):
+
     for i in range(len(hidden_neurons)):
         if hidden_neurons[i] == 0:
             hidden_neurons = 1
@@ -891,50 +881,50 @@ def SAENoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2
     trgt_sparse = np_utils.to_categorical(trgt.astype(int))
 
     # load or create cross validation ids
-    CVO = trnparams.NoveltyDetectionFolds(folder=save_path, n_folds=n_folds, trgt=trgt, dev=dev)
+    CVO = trnparams.NoveltyDetectionFolds(folder=save_path,n_folds=n_folds,trgt=trgt,dev=dev)
 
     n_folds = len(CVO[0])
     n_inits = trn_params.params['n_inits']
-    prefix_str = 'RawData_%i_novelty' % (inovelty)
+    prefix_str = 'RawData_%i_novelty'%(inovelty)
     analysis_str = 'StackedAutoEncoder'
     params_str = trn_params.get_params_str()
 
     if layer == 1:
         neurons_str = str(data.shape[1]) + 'x' + str(n_neurons)
-        model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                         prefix_str, n_folds,
-                                                         params_str, neurons_str)
+        model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                          prefix_str,n_folds,
+                                                          params_str,neurons_str)
     else:
         neurons_str = str(data.shape[1])
         for ineuron in hidden_neurons[:layer]:
             neurons_str = neurons_str + 'x' + str(ineuron)
         neurons_str = neurons_str + 'x' + str(n_neurons)
-        model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                         prefix_str, n_folds,
-                                                         params_str, neurons_str)
+        model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                prefix_str,n_folds,
+                                                                params_str,neurons_str)
 
     if not dev:
-        file_name = '%s_fold_%i_model.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model.h5'%(model_str,ifold)
         if os.path.exists(file_name):
             if trn_params.params['verbose']:
-                print 'File %s exists' % (file_name)
+                print 'File %s exists'%(file_name)
             return 0
     else:
-        file_name = '%s_fold_%i_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model_dev.h5'%(model_str,ifold)
         if os.path.exists(file_name):
             if trn_params.params['verbose']:
-                print 'File %s exists' % (file_name)
+                print 'File %s exists'%(file_name)
             return 0
 
     train_id, test_id = CVO[inovelty][ifold]
 
-    # normalize data based in train set
+    #normalize data based in train set
     if trn_params.params['norm'] == 'mapstd':
-        scaler = preprocessing.StandardScaler().fit(data[train_id, :])
+        scaler = preprocessing.StandardScaler().fit(data[train_id,:])
     elif trn_params.params['norm'] == 'mapstd_rob':
-        scaler = preprocessing.RobustScaler().fit(data[train_id, :])
+        scaler = preprocessing.RobustScaler().fit(data[train_id,:])
     elif trn_params.params['norm'] == 'mapminmax':
-        scaler = preprocessing.MinMaxScaler().fit(data[train_id, :])
+        scaler = preprocessing.MinMaxScaler().fit(data[train_id,:])
 
     norm_data = scaler.transform(data)
     norm_trgt = norm_data
@@ -946,11 +936,11 @@ def SAENoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2
     trn_desc = {}
 
     for i_init in range(n_inits):
-        print 'Neuron: %i - Fold %i of %i Folds -  Init %i of %i Inits' % (n_neurons,
-                                                                           ifold + 1,
-                                                                           n_folds,
-                                                                           i_init + 1,
-                                                                           n_inits)
+        print 'Neuron: %i - Fold %i of %i Folds -  Init %i of %i Inits'%(n_neurons,
+                                                                         ifold+1,
+                                                                         n_folds,
+                                                                         i_init+1,
+                                                                         n_inits)
 
         if layer == 1:
             model = Sequential()
@@ -963,17 +953,17 @@ def SAENoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2
         if layer == 2:
             # Load second layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 1]:
+            for ineuron in hidden_neurons[:layer-1]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             first_layer_model = load_model(file_name)
 
@@ -994,43 +984,43 @@ def SAENoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2
         if layer == 3:
             # Load first layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 2]:
+            for ineuron in hidden_neurons[:layer-2]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             first_layer_model = load_model(file_name)
 
             get_first_layer_output = K.function([first_layer_model.layers[0].input],
-                                                [first_layer_model.layers[1].output])
+                                          [first_layer_model.layers[1].output])
 
             # Projection of first layer
             first_proj_all_data = get_first_layer_output([norm_data])[0]
             # Load second layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 1]:
+            for ineuron in hidden_neurons[:layer-1]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             second_layer_model = load_model(file_name)
 
             get_second_layer_output = K.function([second_layer_model.layers[0].input],
-                                                 [second_layer_model.layers[1].output])
+                                          [second_layer_model.layers[1].output])
 
             # Projection of second layer
             second_proj_all_data = get_second_layer_output([first_proj_all_data])[0]
@@ -1046,66 +1036,66 @@ def SAENoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2
         if layer == 4:
             # Load first layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 3]:
+            for ineuron in hidden_neurons[:layer-3]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             first_layer_model = load_model(file_name)
 
             get_first_layer_output = K.function([first_layer_model.layers[0].input],
-                                                [first_layer_model.layers[1].output])
+                                          [first_layer_model.layers[1].output])
 
             # Projection of first layer
             first_proj_all_data = get_first_layer_output([norm_data])[0]
 
             # Load second layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 2]:
+            for ineuron in hidden_neurons[:layer-2]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             second_layer_model = load_model(file_name)
 
             get_second_layer_output = K.function([second_layer_model.layers[0].input],
-                                                 [second_layer_model.layers[1].output])
+                                          [second_layer_model.layers[1].output])
 
             # Projection of second layer
             second_proj_all_data = get_second_layer_output([first_proj_all_data])[0]
 
             # Load third layer model
             neurons_str = str(data.shape[1])
-            for ineuron in hidden_neurons[:layer - 1]:
+            for ineuron in hidden_neurons[:layer-1]:
                 neurons_str = neurons_str + 'x' + str(ineuron)
-            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                                      prefix_str,
-                                                                      n_folds,
-                                                                      params_str,
-                                                                      neurons_str)
+            previous_model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                                    prefix_str,
+                                                                    n_folds,
+                                                                    params_str,
+                                                                    neurons_str)
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str,ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str,ifold)
 
             third_layer_model = load_model(file_name)
 
             get_third_layer_output = K.function([third_layer_model.layers[0].input],
-                                                [third_layer_model.layers[1].output])
+                                          [third_layer_model.layers[1].output])
 
             # Projection of third layer
             third_proj_all_data = get_third_layer_output([second_proj_all_data])[0]
@@ -1117,7 +1107,7 @@ def SAENoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2
             model.add(Activation(trn_params.params['output_activation']))
 
             norm_data = third_proj_all_data
-        # end if layer == 4
+        #end if layer == 4
 
         # Optmizer
         adam = Adam(lr=trn_params.params['learning_rate'],
@@ -1154,30 +1144,30 @@ def SAENoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2
 
     # save model
     if not dev:
-        file_name = '%s_fold_%i_model.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model.h5'%(model_str,ifold)
         classifier.save(file_name)
-        file_name = '%s_fold_%i_trn_desc.jbl' % (model_str, ifold)
-        joblib.dump([trn_desc], file_name, compress=9)
+        file_name = '%s_fold_%i_trn_desc.jbl'%(model_str,ifold)
+        joblib.dump([trn_desc],file_name,compress=9)
     else:
-        file_name = '%s_fold_%i_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_fold_%i_model_dev.h5'%(model_str,ifold)
         classifier.save(file_name)
-        file_name = '%s_fold_%i_trn_desc_dev.jbl' % (model_str, ifold)
-        joblib.dump([trn_desc], file_name, compress=9)
+        file_name = '%s_fold_%i_trn_desc_dev.jbl'%(model_str,ifold)
+        joblib.dump([trn_desc],file_name,compress=9)
 
     return [classifier, trn_desc]
 
-
-def SAEClassifierNoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2, hidden_neurons=[400],
+def SAEClassifierNoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0, n_folds=2, hidden_neurons = [400],
                                       trn_params=None, save_path='', verbose=False, dev=False):
+
     # turn targets in sparse mode
     trgt_sparse = np_utils.to_categorical(trgt.astype(int))
 
     # load or create cross validation ids
-    CVO = trnparams.NoveltyDetectionFolds(folder=save_path, n_folds=n_folds, trgt=trgt, dev=dev)
+    CVO = trnparams.NoveltyDetectionFolds(folder=save_path,n_folds=n_folds,trgt=trgt,dev=dev)
 
     n_folds = len(CVO[inovelty])
     n_inits = trn_params.params['n_inits']
-    prefix_str = 'RawData_%i_novelty' % (inovelty)
+    prefix_str = 'RawData_%i_novelty'%(inovelty)
     analysis_str = 'StackedAutoEncoder'
     params_str = trn_params.get_params_str()
 
@@ -1185,25 +1175,25 @@ def SAEClassifierNoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0,
     for ineuron in hidden_neurons:
         neurons_str = neurons_str + 'x' + str(ineuron)
 
-    model_str = '%s/%s/%s_%i_folds_%s_%s_neurons' % (save_path, analysis_str,
-                                                     prefix_str, n_folds,
-                                                     params_str, neurons_str)
+    model_str = '%s/%s/%s_%i_folds_%s_%s_neurons'%(save_path,analysis_str,
+                                                   prefix_str,n_folds,
+                                                   params_str,neurons_str)
 
     if not dev:
-        file_name = '%s_classifier_fold_%i_model.h5' % (model_str, ifold)
+        file_name = '%s_classifier_fold_%i_model.h5'%(model_str,ifold)
     else:
-        file_name = '%s_classifier_fold_%i_model_dev.h5' % (model_str, ifold)
+        file_name = '%s_classifier_fold_%i_model_dev.h5'%(model_str,ifold)
 
     if not os.path.exists(file_name):
         train_id, test_id = CVO[inovelty][ifold]
 
-        # normalize data based in train set
+        #normalize data based in train set
         if trn_params.params['norm'] == 'mapstd':
-            scaler = preprocessing.StandardScaler().fit(data[train_id, :])
+            scaler = preprocessing.StandardScaler().fit(data[train_id,:])
         elif trn_params.params['norm'] == 'mapstd_rob':
-            scaler = preprocessing.RobustScaler().fit(data[train_id, :])
+            scaler = preprocessing.RobustScaler().fit(data[train_id,:])
         elif trn_params.params['norm'] == 'mapminmax':
-            scaler = preprocessing.MinMaxScaler().fit(data[train_id, :])
+            scaler = preprocessing.MinMaxScaler().fit(data[train_id,:])
 
         norm_data = scaler.transform(data)
         norm_trgt = norm_data
@@ -1215,24 +1205,25 @@ def SAEClassifierNoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0,
         trn_desc = {}
 
         for i_init in range(n_inits):
-            print 'Topology: %sx%i - Fold %i of %i Folds -  Init %i of %i Inits' % (neurons_str,
-                                                                                    trgt_sparse.shape[1],
-                                                                                    ifold + 1,
-                                                                                    n_folds,
-                                                                                    i_init + 1,
-                                                                                    n_inits)
+            print 'Topology: %sx%i - Fold %i of %i Folds -  Init %i of %i Inits'%(neurons_str,
+                                                                                  trgt_sparse.shape[1],
+                                                                                  ifold+1,
+                                                                                  n_folds,
+                                                                                  i_init+1,
+                                                                                  n_inits)
+
 
             # First Layer
-            previous_model_str = '%s/%s/%s_%i_folds_%s_400x%i_neurons' % (save_path, analysis_str,
-                                                                          prefix_str,
-                                                                          n_folds,
-                                                                          params_str,
-                                                                          hidden_neurons[0])
+            previous_model_str = '%s/%s/%s_%i_folds_%s_400x%i_neurons'%(save_path,analysis_str,
+                                                                        prefix_str,
+                                                                        n_folds,
+                                                                        params_str,
+                                                                        hidden_neurons[0])
 
             if not dev:
-                file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model.h5'%(previous_model_str, ifold)
             else:
-                file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str, ifold)
 
             # Get first layer weights
             first_layer_model = load_model(file_name)
@@ -1241,23 +1232,22 @@ def SAEClassifierNoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0,
             model = Sequential()
 
             # Encoder of first layer
-            model.add(
-                Dense(hidden_neurons[0], input_dim=norm_data.shape[1], weights=encoder_first_layer, trainable=False))
+            model.add(Dense(hidden_neurons[0], input_dim=norm_data.shape[1], weights=encoder_first_layer, trainable=False))
             model.add(Activation(trn_params.params['hidden_activation']))
 
             if len(hidden_neurons) > 1:
                 # Second Layer
-                previous_model_str = '%s/%s/%s_%i_folds_%s_400x%ix%i_neurons' % (save_path, analysis_str,
-                                                                                 prefix_str,
-                                                                                 n_folds,
-                                                                                 params_str,
-                                                                                 hidden_neurons[0],
-                                                                                 hidden_neurons[1])
+                previous_model_str = '%s/%s/%s_%i_folds_%s_400x%ix%i_neurons'%(save_path,analysis_str,
+                                                                        prefix_str,
+                                                                        n_folds,
+                                                                        params_str,
+                                                                        hidden_neurons[0],
+                                                                        hidden_neurons[1])
 
                 if not dev:
-                    file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                    file_name = '%s_fold_%i_model.h5'%(previous_model_str, ifold)
                 else:
-                    file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                    file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str, ifold)
                 # Get the second layer projection of data
                 second_layer_model = load_model(file_name)
                 encoder_second_layer = second_layer_model.layers[0].get_weights()
@@ -1268,18 +1258,18 @@ def SAEClassifierNoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0,
 
             if len(hidden_neurons) > 2:
                 # Third Layer
-                previous_model_str = '%s/%s/%s_%i_folds_%s_400x%ix%ix%i_neurons' % (save_path, analysis_str,
-                                                                                    prefix_str,
-                                                                                    n_folds,
-                                                                                    params_str,
-                                                                                    hidden_neurons[0],
-                                                                                    hidden_neurons[1],
-                                                                                    hidden_neurons[2])
+                previous_model_str = '%s/%s/%s_%i_folds_%s_400x%ix%ix%i_neurons'%(save_path,analysis_str,
+                                                                        prefix_str,
+                                                                        n_folds,
+                                                                        params_str,
+                                                                        hidden_neurons[0],
+                                                                        hidden_neurons[1],
+                                                                        hidden_neurons[2])
 
                 if not dev:
-                    file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                    file_name = '%s_fold_%i_model.h5'%(previous_model_str, ifold)
                 else:
-                    file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                    file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str, ifold)
 
                 # Get the third layer projection of data
                 third_layer_model = load_model(file_name)
@@ -1291,19 +1281,19 @@ def SAEClassifierNoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0,
 
             if len(hidden_neurons) > 3:
                 # Fourth Layer
-                previous_model_str = '%s/%s/%s_%i_folds_%s_400x%ix%ix%ix%i_neurons' % (save_path, analysis_str,
-                                                                                       prefix_str,
-                                                                                       n_folds,
-                                                                                       params_str,
-                                                                                       hidden_neurons[0],
-                                                                                       hidden_neurons[1],
-                                                                                       hidden_neurons[2],
-                                                                                       hidden_neurons[3])
+                previous_model_str = '%s/%s/%s_%i_folds_%s_400x%ix%ix%ix%i_neurons'%(save_path,analysis_str,
+                                                                        prefix_str,
+                                                                        n_folds,
+                                                                        params_str,
+                                                                        hidden_neurons[0],
+                                                                        hidden_neurons[1],
+                                                                        hidden_neurons[2],
+                                                                        hidden_neurons[3])
 
                 if not dev:
-                    file_name = '%s_fold_%i_model.h5' % (previous_model_str, ifold)
+                    file_name = '%s_fold_%i_model.h5'%(previous_model_str, ifold)
                 else:
-                    file_name = '%s_fold_%i_model_dev.h5' % (previous_model_str, ifold)
+                    file_name = '%s_fold_%i_model_dev.h5'%(previous_model_str, ifold)
 
                 # Get the third layer projection of data
                 fourth_layer_model = load_model(file_name)
@@ -1352,34 +1342,33 @@ def SAEClassifierNoveltyTrainFunction(data=None, trgt=None, inovelty=0, ifold=0,
 
         # save model
         if not dev:
-            file_name = '%s_classifier_fold_%i_model.h5' % (model_str, ifold)
+            file_name = '%s_classifier_fold_%i_model.h5'%(model_str,ifold)
             classifier.save(file_name)
-            file_name = '%s_classifier_fold_%i_trn_desc.jbl' % (model_str, ifold)
-            joblib.dump([trn_desc], file_name, compress=9)
+            file_name = '%s_classifier_fold_%i_trn_desc.jbl'%(model_str,ifold)
+            joblib.dump([trn_desc],file_name,compress=9)
         else:
-            file_name = '%s_classifier_fold_%i_model_dev.h5' % (model_str, ifold)
+            file_name = '%s_classifier_fold_%i_model_dev.h5'%(model_str,ifold)
             classifier.save(file_name)
-            file_name = '%s_classifier_fold_%i_trn_desc_dev.jbl' % (model_str, ifold)
-            joblib.dump([trn_desc], file_name, compress=9)
+            file_name = '%s_classifier_fold_%i_trn_desc_dev.jbl'%(model_str,ifold)
+            joblib.dump([trn_desc],file_name,compress=9)
     else:
         # The model has already been trained, so load the files
         if verbose:
             print 'Load Model'
         classifier = Sequential()
         if not dev:
-            file_name = '%s_classifier_fold_%i_model.h5' % (model_str, ifold)
+            file_name = '%s_classifier_fold_%i_model.h5'%(model_str,ifold)
         else:
-            file_name = '%s_classifier_fold_%i_model_dev.h5' % (model_str, ifold)
+            file_name = '%s_classifier_fold_%i_model_dev.h5'%(model_str,ifold)
         classifier = load_model(file_name)
 
         if not dev:
-            file_name = '%s_classifier_fold_%i_trn_desc.jbl' % (model_str, ifold)
+            file_name = '%s_classifier_fold_%i_trn_desc.jbl'%(model_str,ifold)
         else:
-            file_name = '%s_classifier_fold_%i_trn_desc_dev.jbl' % (model_str, ifold)
+            file_name = '%s_classifier_fold_%i_trn_desc_dev.jbl'%(model_str,ifold)
         [trn_desc] = joblib.load(file_name)
 
-    return [classifier, trn_desc]
-
+    return [classifier,trn_desc]
 
 class ConvolutionTrainFunction(ConvolutionPaths):
     def __init__(self):
@@ -1397,11 +1386,6 @@ class ConvolutionTrainFunction(ConvolutionPaths):
     def loadModels(self, model_list, model_initializer):
         self.models = [model_initializer(model) for model in model_list]
 
-    # def loadFolds(self, n_folds, mode):
-    #     file_name = '%s/%i_folds_%s.jbl' % (self.results_path, n_folds, mode)
-    #     print "Loading fold configuration"
-    #     self.fold_config = SystemIO.load(file_name)
-    #     self.n_folds = n_folds
     def loadFolds(self, folds):
         self.fold_config = folds
         self.n_folds = len(folds)
@@ -1495,8 +1479,8 @@ class ConvolutionTrainFunction(ConvolutionPaths):
 
             class_weights = self._getGradientWeights(y_train, mode='standard', novelty_cls=novelty_cls)
 
-            y_test = DataHandler.trgt2categorical(y_test, n_classes)
-            y_train = DataHandler.trgt2categorical(y_train, n_classes)
+            y_test = Functions.NpUtils.DataTransformation.trgt2categorical(y_test, n_classes)
+            y_train = Functions.NpUtils.DataTransformation.trgt2categorical(y_train, n_classes)
 
             nv_mask = np.ones(y_train.shape[1], dtype=bool)
             #nv_mask[novelty_cls or []] = False
@@ -1546,12 +1530,10 @@ class ConvolutionTrainFunction(ConvolutionPaths):
         os.remove(model.model_recovery_history)
         return model_history, model_predictions
 
-
     @staticmethod
     def _parseFoldNumber(fold_str):
         warn("Limit : [0,9]!")
         return int(fold_str[0])
-
 
     @staticmethod
     def _saveResults(model, model_results, novelty_classes, class_labels):
@@ -1576,7 +1558,6 @@ class ConvolutionTrainFunction(ConvolutionPaths):
         pd_pred.to_csv(preds_file, sep=',')
         pd_hist.to_csv(hist_file, sep=',')
 
-
     def _getGradientWeights(self, y_train, mode = 'standard', novelty_cls=None):
         cls_indices, event_count = np.unique(np.array(y_train), return_counts=True)
         min_class = min(event_count)
@@ -1586,141 +1567,3 @@ class ConvolutionTrainFunction(ConvolutionPaths):
         print cls_indices
         return {cls_index: float(min_class) / cls_count
                 for cls_index, cls_count in zip(cls_indices, event_count)}
-
-
-class _CNNModel(ModelPaths):
-    def __init__(self, trnParams):
-        super(_CNNModel, self).__init__(trnParams)
-        if not isinstance(trnParams, TrnParamsConvolutional):
-            raise ValueError('The parameters inserted must be an instance of TrnPararamsCNN'
-                             '%s of type %s was passed.'
-                             % (trnParams, type(trnParams)))
-
-        self.model = None
-        self.params = None
-
-        self.path = ModelPaths(trnParams)
-        if SystemIO.exists(self.path.model_path):
-            trnParams = self.loadInfo()  # see __setstate__ and arb obj storage with joblib
-            self.mountParams(trnParams)
-        else:
-            self.createFolders(self.results_path + '/' + trnParams.getParamPath())
-            self.mountParams(trnParams)
-            self.saveParams(trnParams)
-
-    def summary(self):
-        raise NotImplementedError
-
-    def mountParams(self, trnParams):
-        for param, value in trnParams:
-            if param is None:
-                raise ValueError('The parameters configuration received by _CNNModel must be complete'
-                                 '%s was passed as NoneType.'
-                                 % (param))
-            setattr(self, param, value)
-
-        # self.model_name = trnParams.getParamStr()
-
-    def saveParams(self, trnParams):
-        SystemIO.save(trnParams.toNpArray(), self.path.model_info_file)
-
-    def loadInfo(self):
-        params_array = SystemIO.load(self.path.model_info_file)
-        return TrnParamsConvolutional(*params_array)
-
-    @abstractmethod
-    def load(self, file_path):
-        pass
-
-    @abstractmethod
-    def save(self, filepath):
-        pass
-
-    @abstractmethod
-    def build(self):
-        pass
-
-    @abstractmethod
-    def fit(self, x_train, y_train, validation_data=None, class_weight=None, verbose=0):
-        pass
-
-    @abstractmethod
-    def evaluate(self, x_test, y_test, verbose=0):
-        pass
-
-    @abstractmethod
-    def predict(self, data, verbose=0):
-        pass
-
-
-class KerasModel(_CNNModel):
-    def __init__(self, trnParams):
-        super(KerasModel, self).__init__(trnParams)
-
-    def build(self):
-        if self.model is not None:
-            warnings.warn('Model is not empty and was already trained.\n'
-                          'Run purge method for deleting the model variable',
-                          Warning)
-
-        self.purge()
-        self.model = Sequential()
-        for layer in self.layers:
-            self.model.add(layer.toKerasFn())
-
-        self.model.compile(optimizer=self.optimizer.toKerasFn(),
-                           loss=self.loss,
-                           metrics=self.metrics
-                           )
-
-    def fit(self, x_train, y_train, callbacks, validation_data=None, class_weight=None, verbose=0):
-        if self.model is None:
-            raise StandardError('Model is not built. Run build method or load model before fitting')
-
-        print self.epochs
-        print self.batch_size
-        print self.callbacks
-        print class_weight
-
-        history = self.model.fit(x_train,
-                                 y_train,
-                                 epochs=self.epochs,
-                                 batch_size=self.batch_size,
-                                 callbacks=callbacks,
-                                 validation_data=validation_data,
-                                 class_weight=class_weight,
-                                 verbose=verbose
-                                 )
-        self.history = history.history
-        return history.history
-
-    def evaluate(self, x_test, y_test, verbose=0):
-        if self.model is None:
-            raise StandardError('Model is not built. Run build method or load model before fitting')
-
-        val_history = self.model.evaluate(x_test,
-                                          y_test,
-                                          batch_size=self.batch_size,
-                                          verbose=verbose)
-        self.val_history = val_history
-        return val_history
-
-    def get_layer_n_output(self, n, data):
-        intermediate_layer_model = Model(inputs=self.model.input,
-                                         outputs=self.model.layers[n - 1].output)
-        intermediate_output = intermediate_layer_model.predict(data)
-        return intermediate_output
-
-    def predict(self, data, verbose=0):
-        if self.model is None:
-            raise StandardError('Model is not built. Run build method or load model before fitting')
-        return self.model.predict(data, 1, verbose)  # ,steps)
-
-    def load(self, file_path):
-        self.model = load_model(file_path)
-
-    def save(self, file_path):
-        self.model.save(file_path)
-
-    def purge(self):
-        del self.model
